change integer to 64 bit

+++ WheyVM/Operand/Types.h
@@ -6,7 +6,7 @@
 #define BOOLEAN_TRUE 1
 #define BOOLEAN_FALSE 0
 
-typedef int32_t Integer;
+typedef int64_t Integer;
 typedef double Double;
 typedef char Char;
